android: move base adapter to lib
