+++ android/Lib/src/test/kotlin/com/darakeon/dfm/lib/ui/MockAdapter.kt
@@ -1,4 +1,4 @@
-package com.darakeon.dfm.base
+package com.darakeon.dfm.lib.ui
 
 import android.app.Activity
 import android.view.View
@@ -9,7 +9,7 @@ class MockAdapter(
 	list: List<String>
 ): Adapter<Activity, String, View>(activity, list)
 {
-	override val id = android.R.layout.simple_list_item_1
+	override val lineLayoutId = android.R.layout.simple_list_item_1
 	override fun populateView(view: View, position: Int) {
 		view.findViewById<TextView>(
 			android.R.id.text1
